Fix spin count issue for values less than 60
- Changed shoes_per_session from int to float in all three locations
- Updated StrategyOverrides type annotation to float
- Now supports fractional shoe values (0.5 = 30 spins, 0.75 = 45 spins, etc.)
- Previously int(0.5) = 0 caused sessions to have 0 spins
- Added test to verify fix works for various spin counts

+++ engine/strategy_rules.py
@@ -29,7 +29,7 @@ class StrategyOverrides:
     bet_strategy_2: str = None 
     
     # --- SESSION ---
-    shoes_per_session: int = 3
+    shoes_per_session: float = 3.0
     penalty_box_enabled: bool = True
     
     # --- BACCARAT TIE BETTING ---
